开源硬件大作业/requirements.txt
pyaudio
mcpi
keyboard
baidu-aip
PyUserInput
chardet
pynput
mediapipe
numpy